Build fix.
svn:sync-xref-src-repo-rev: r124505

+++ src/VBox/Devices/Audio/AudioMixer.cpp
@@ -1,4 +1,4 @@
-/* $Id: AudioMixer.cpp 73813 2018-08-22 08:39:00Z andreas.loeffler@oracle.com $ */
+/* $Id: AudioMixer.cpp 73815 2018-08-22 08:43:28Z andreas.loeffler@oracle.com $ */
 /** @file
  * Audio mixing routines for multiplexing audio sources in device emulations.
  *
@@ -1914,7 +1914,7 @@ static int audioMixerSinkMultiplexSync(PAUDMIXSINK pSink, AUDMIXOP enmOp, const 
         if (!DrvAudioHlpStreamStatusCanWrite(pMixStream->pConn->pfnStreamGetStatus(pMixStream->pConn, pMixStream->pStream)))
             continue;
 
-        cbToWriteMin = RT_MIN(cbBuf, RT_MIN(cbToWriteMin, RTCircBufFree(pMixStream->pCircBuf)));
+        cbToWriteMin = RT_MIN(cbBuf, RT_MIN(cbToWriteMin, (uint32_t)RTCircBufFree(pMixStream->pCircBuf)));
     }
 
     if (cbToWriteMin == UINT32_MAX) /* No space at all? */
